backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# NLP & ML (Optional - for MVP, using simpler approach)
# spacy==3.7.2  # Uncomment if you need advanced NLP
# sentence-transformers==2.3.1  # Uncomment for semantic similarity
# transformers==4.37.0  # Uncomment for transformer models
# torch==2.2.0  # Uncomment for GPU support (CPU mode below)
numpy==1.26.3
scikit-learn==1.4.0

# Vector Database (Optional for MVP)
# chromadb==0.4.22

# AI APIs
openai==1.10.0
anthropic==0.18.1

# Citation Processing
citeproc-py==0.6.0
pybtex==0.24.0

# HTTP & External APIs
httpx==0.26.0
aiohttp==3.9.1

# Background Tasks
celery==5.3.6
flower==2.0.1

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.2
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Monitoring
sentry-sdk[fastapi]==1.40.0